renamed to state_trigger_panel
